Move unit.isAliveTrick to widget.isAliveTrick
The trick applies to widgets in general and not just units.

Note that there is a technical difference in the implementation:
unit.isAliveTrick used the GetunitState native whereas widget.isAliveTrick
uses the GetWidgetLife native. I'm unaware if there are any differences
in the output of said native functions. I assume not.