fix ClosureForGroups.forNearestDestructable() (#290)
* fix ClosureForGroups.forNearestDestructable()

Previous version was not working it all. It was using filterEnumDestructablesInCircleBJ, which uses some bj vars and requires them to be set, which were not. It could be fixed with just setting this variables, but previous implementation has some other issues: distance was calculated two times (in the filter function and in the enum function) and the rect created was not removed! Now a global rect is reused, and the in-rect-check was moved to enum function to avoid recalculating distance.

* Create test.j

* delete trash

* destroy callback