Add support for iterating over forces.
Iterating over forces requires a `code` right now. In lieu of adding support for closures or converting every force to a list / set, I'm adding these methods so forces can at least be used in a standard for-loop. This is destructive on the original force.

This can be used like:

```
for target in bj_FORCE_ALL_PLAYERS.iterator()
    print(target.getName())
```